refactor: WinningPrizeCalculator 삭제
