+++ src/main/kotlin/lotto/model/LottoStore.kt
@@ -18,8 +18,17 @@ class LottoStore(private val lottoNumberGenerator: NumberGenerator) {
         }
     }
 
-    fun getWinningResult(winningLotto: WinningLotto): Map<WinningRank, Int> {
-        return lottos.winningResult(winningLotto)
+    private fun calculateProfitAmount(winningResult: Map<WinningRank, Int>): Int =
+        winningResult.entries.sumOf {
+            it.key.prize * (it.value)
+        }.toInt()
+
+    fun calculateProfitRate(
+        purchaseAmount: Int,
+        winningResult: Map<WinningRank, Int>,
+    ): Double {
+        val profitAmount = calculateProfitAmount(winningResult)
+        return (profitAmount / purchaseAmount).toDouble()
     }
 
     companion object {
